pyproject.toml
[project]
name = "prompt-eval"
version = "0.1.0"
description = "LLM Evaluation and Grading System"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "langchain>=0.2.5",
    "python-dotenv>=1.2.2",
    "langchain-groq>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "uvicorn>=0.27.0",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
]
